Much documentation. Many changes and cleanup. Added better tests

+++ src/vm.rs
@@ -1,3 +1,8 @@
+//!
+//! Module containing a higher-level wrapper over a raw JVM. This higher-level implementation
+//! can handle automatic creation/destruction, all the Invoke Interface functions, and tries to
+//! ensure safety while doing so.
+//!
 
 use crate::{env, ffi};
 use crate::error::Error;
@@ -8,25 +13,27 @@ use crate::env::JNIEnv;
 /// Higher-level construct representing a JVM
 pub struct JavaVM {
     version: JNIVersion,
-    main_vm: *mut ffi::JavaVM
+    main_vm: *mut ffi::JavaVM,
+    owned: bool
 }
 
 impl JavaVM {
 
     /// Build a JVM instance from a version and pointer
-    pub fn new(version: JNIVersion, vm: *mut ffi::JavaVM) -> Result<JavaVM, Error> {
+    pub fn new(version: JNIVersion, vm: *mut ffi::JavaVM, owned: bool) -> Result<JavaVM, Error> {
         if vm.is_null() {
             Err(Error::new("JavaVM must be constructed from non-null pointer", ffi::constants::JNI_ERR))
         } else {
             Ok(JavaVM {
                 version,
-                main_vm: vm
+                main_vm: vm,
+                owned
             })
         }
     }
 
     /// Create a new JVM. Initializes an entirely new JVM, with the current thread
-    /// as the main thread.
+    /// as the main thread. This object will call the JVM destroy function when it is dropped
     pub fn create(version: JNIVersion) -> Result<(JavaVM, JNIEnv), Error> {
         let mut main_vm = std::ptr::null_mut();
         let mut main_env = std::ptr::null_mut();
@@ -51,7 +58,30 @@ impl JavaVM {
             Err(Error::new("Main VM or Global Environment null, despite successful JVM creation", ffi::constants::JNI_ERR))
         } else {
             let main_env = env::JNIEnv::new(main_env)?;
-            Ok((JavaVM { version, main_vm }, main_env))
+            Ok((JavaVM { version, main_vm, owned: true }, main_env))
+        }
+    }
+
+    /// Get a list of existing JVMs. May error if the JNI returns an error code
+    pub fn get_existing(version: JNIVersion) -> Result<Vec<JavaVM>, Error> {
+        const BASE_QUANT: usize = 8;
+
+        let mut main_vms: [*mut ffi::JavaVM; BASE_QUANT] = [std::ptr::null_mut(); BASE_QUANT];
+        let mut total: i32 = 0;
+
+        let result = unsafe {
+            ffi::get_created_jvms(main_vms.as_mut_ptr(), BASE_QUANT as _, &mut total)
+        };
+
+        if result != 0 {
+            Err(Error::new("Couldn't get list of existing JVM instances", result))
+        } else {
+            let mut out = Vec::new();
+            for idx in 0..total as usize {
+                out.push(JavaVM::new(version, main_vms[idx], false))
+            }
+
+            out.into_iter().collect()
         }
     }
 
@@ -126,10 +156,12 @@ impl JavaVM {
 
 impl Drop for JavaVM {
     fn drop(&mut self) {
-        let vm = self.internal_vm();
-        let result = vm.destroy_java_vm();
-        if result != 0 {
-            panic!(format!("JVM failed to shut down: {}", result));
+        if self.owned {
+            let vm = self.internal_vm();
+            let result = vm.destroy_java_vm();
+            if result != 0 {
+                panic!(format!("JVM failed to shut down: {}", result));
+            }
         }
     }
 }
@@ -141,6 +173,7 @@ unsafe impl Sync for JavaVM {}
 mod tests {
     use super::*;
     use rust_jni_proc::java;
+    use crate::tests::with_vm;
     use crate::types::{JavaDownCast, JObject, JClass, JString};
     use crate::env::JNIEnv;
     use crate::JThrowable;
@@ -153,7 +186,7 @@ mod tests {
     }
 
     #[java(class = "java.lang.Foo", name = "Foo")]
-    fn func(env: &JNIEnv, obj: JObject, arg1: JClass, arg2: Option<JString>) -> JObject {
+    fn func(env: &JNIEnv, _this: JObject, _arg1: JClass, _arg2: Option<JString>) -> JObject {
         JObject::new(std::ptr::null_mut()).unwrap()
     }
 
@@ -163,12 +196,12 @@ mod tests {
     }
 
     #[java(class = "temp.foo.Foo")]
-    fn func(env: &JNIEnv, obj: JObject) -> Result<JObject, Error> {
+    fn func(env: &JNIEnv, _this: JObject) -> Result<JObject, Error> {
         Err(Error::new("", -1))
     }
 
     #[java(class = "temp.bar.Bar")]
-    fn func(env: &JNIEnv, obj: JObject) -> Result<JObject, JThrowable> {
+    fn func(env: &JNIEnv, _this: JObject) -> Result<JObject, JThrowable> {
         Ok(JObject::new(std::ptr::null_mut()).unwrap())
     }
 
@@ -180,42 +213,34 @@ mod tests {
 
     #[test]
     fn test() {
-        let (vm, env) = JavaVM::create(JNIVersion::Ver18).unwrap();
+        with_vm(|vm| {
+            let env = vm.attach_current_thread().expect("Couldn't attach test thread");
 
-        check_cls(&env, "java.lang.String");
-        check_cls(&env, "java.lang.String[]");
-        check_cls(&env, "int[]");
-        let cls = env.find_class("java.lang.Class").unwrap();
-        let method_id = env.get_static_method_id(&cls, "getPrimitiveClass", "(java.lang.String) -> java.lang.Class").unwrap();
+            let cls = env.find_class("java.lang.Class").unwrap();
+            let method_id = env.get_static_method_id(&cls, "getPrimitiveClass", "(java.lang.String) -> java.lang.Class").unwrap();
 
-        let str = env.new_string_utf("void").unwrap();
-        let args = vec![str.downcast().into()];
-        let result = env.call_static_method(
-            &cls,
-            &method_id,
-            &args
-        ).unwrap().unwrap();
+            let str = env.new_string_utf("void").unwrap();
+            let args = vec![str.downcast().into()];
+            env.call_static_method(
+                &cls,
+                &method_id,
+                &args
+            ).unwrap().unwrap();
 
-        println!("{:?}", result);
+            use crate::{get_cls, get_method_id};
 
-        use crate::{get_cls, get_method_id, get_static_method_id, get_field_id, get_static_field_id};
+            let cls = get_cls!(env, "java.lang.String");
+            let _id = get_method_id!(env, cls, "length", "() -> int");
+            // let _id = get_static_method_id!(env, cls, "blah", "() -> void");
+            // let _id = get_field_id!(env, cls, "field", "java.lang.String");
+            // let _id = get_static_field_id!(env, cls, "field", "java.lang.String");
 
-        let cls = get_cls!(env, "java.lang.String");
-        let _id = get_method_id!(env, cls, "length", "() -> int");
-        let _id = get_static_method_id!(env, cls, "blah", "() -> void");
-        let _id = get_field_id!(env, cls, "field", "java.lang.String");
-        let _id = get_static_field_id!(env, cls, "field", "java.lang.String");
-
-        let obj_buff;
-        {
-            let mut buff = [0x1, 0x2, 0x3];
-            obj_buff = env.new_direct_byte_buffer(&mut buff).unwrap();
-            env.get_direct_buffer_slice(&obj_buff).unwrap();
-        }
-
-        std::mem::drop(vm);
-
-        // let _env2 = vm.get_local_env().unwrap();
-        // let _vm = JavaVM::create(JNIVersion::Ver18).unwrap();
+            let obj_buff;
+            {
+                let mut buff = [0x1, 0x2, 0x3];
+                obj_buff = env.new_direct_byte_buffer(&mut buff).unwrap();
+                env.get_direct_buffer_slice(&obj_buff).unwrap();
+            }
+        });
     }
 }
